# matrix_actions_api/gabe-matrix-actions-openapi.json
{"openapi":"3.1.0","info":{"title":"Gabe Matrix Actions API","description":"HTTPS bridge that allows ChatGPT GPT Actions to interact with Gabe's Clearing","version":"1.0.0"},"paths":{"/send_message":{"post":{"summary":"Send Message","description":"Send a text event to a Matrix room when explicitly asked.","operationId":"send_message_send_message_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/read_messages":{"get":{"summary":"Read Messages","description":"Return the latest Matrix text events for the requested room.","operationId":"read_messages_read_messages_get","parameters":[{"name":"room_id","in":"query","required":true,"schema":{"type":"string","description":"Matrix room_id to read","title":"Room Id"},"description":"Matrix room_id to read"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Maximum number of events to return","default":20,"title":"Limit"},"description":"Maximum number of events to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MatrixMessage"},"title":"Response Read Messages Read Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/room_members":{"get":{"summary":"Room Members","description":"Return the joined members for the requested Matrix room.","operationId":"room_members_room_members_get","parameters":[{"name":"room_id","in":"query","required":true,"schema":{"type":"string","description":"Matrix room_id to inspect","title":"Room Id"},"description":"Matrix room_id to inspect"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MatrixMember"},"title":"Response Room Members Room Members Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MatrixMember":{"properties":{"user_id":{"type":"string","title":"User Id"},"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"}},"type":"object","required":["user_id","display_name"],"title":"MatrixMember","description":"Member of a Matrix room."},"MatrixMessage":{"properties":{"body":{"type":"string","title":"Body"},"event_id":{"type":"string","title":"Event Id"},"sender":{"type":"string","title":"Sender"},"timestamp":{"type":"integer","title":"Timestamp"}},"type":"object","required":["body","event_id","sender","timestamp"],"title":"MatrixMessage","description":"Normalized Matrix text event returned to GPT."},"SendMessageRequest":{"properties":{"room_id":{"type":"string","title":"Room Id","description":"Matrix room_id provided by the operator"},"message":{"type":"string","minLength":1,"title":"Message","description":"Plain text body to send"}},"type":"object","required":["room_id","message"],"title":"SendMessageRequest","description":"Input payload for the send_message endpoint."},"SendMessageResponse":{"properties":{"event_id":{"type":"string","title":"Event Id"},"room_id":{"type":"string","title":"Room Id"},"status":{"type":"string","title":"Status"}},"type":"object","required":["event_id","room_id","status"],"title":"SendMessageResponse","description":"Response returned after sending a Matrix event."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}